Request for a confirmation before deleting account

+++ client/src/components/the_rest/dashboard/Dashboard.js
@@ -21,8 +21,9 @@ class Dashboard extends React.Component {
     }
 
     deleteProfileFunction() {
-        console.log('ready');
-        this.props.deleteProfile(this.props.history);
+        if (window.confirm('Are you sure? (no way to undo that)')) {
+          this.props.deleteProfile(this.props.history);
+        }
     }
 
     render(){
